parser-produit/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pdfminer.six==20221105
PyPDF2==3.0.1
beautifulsoup4==4.12.2
pytesseract==0.3.10
Pillow==10.1.0
pyzbar==0.1.9
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
